# Misc pyxl2 perf improvements continued