--- a/from_cpython/Include/object.h
+++ b/from_cpython/Include/object.h
@@ -456,7 +456,7 @@ struct _typeobject {
 
     void* _hcls;
     void* _hcattrs;
-    char _ics[32];
+    char _ics[40];
     void* _gcvisit_func;
     int _attrs_offset;
     char _flags[7]; // These are bools in C++