```
                                   upstream/master:      origin/pyxl2_2:
           django_template3.py             2.2s (4)             2.2s (2)  +0.7%
                 pyxl_bench.py             1.8s (4)             1.8s (2)  +0.4%
     sqlalchemy_imperative2.py             2.3s (4)             2.3s (2)  -0.2%
                pyxl_bench2.py             1.3s (4)             1.3s (2)  -1.3%
       django_template3_10x.py            14.2s (4)            14.4s (2)  +0.9%
             pyxl_bench_10x.py            14.3s (4)            14.2s (2)  -0.4%
 sqlalchemy_imperative2_10x.py            17.5s (4)            17.4s (2)  -0.5%
            pyxl_bench2_10x.py            12.2s (4)            11.5s (2)  -5.5%
                       geomean                 5.2s                 5.1s  -0.8%
```